move edge types into models/EdgeData
