+++ Composer/packages/extensions/visual-designer/src/components/lib/EdgeUtil.tsx
@@ -3,39 +3,7 @@
 
 import React from 'react';
 
-interface Coord2D {
-  x: number;
-  y: number;
-}
-
-export enum EdgeDirection {
-  Up = 'Up',
-  Down = 'Down',
-  Left = 'Left',
-  Right = 'Right',
-}
-
-interface EdgeOptions {
-  color?: string;
-
-  /** Indicates if the line stroke is dashed */
-  dashed?: boolean;
-
-  /** If set to true, an arrow will be drawn at the end of the line */
-  arrowed?: boolean;
-
-  /** Text displayed on the edge */
-  label?: string | number;
-
-  /** label text's offset compared with start point's position */
-  labelOptions?: {
-    offset: {
-      x: number;
-      y: number;
-    };
-    fontSize: number;
-  };
-}
+import { Coord2D, EdgeDirection, EdgeOptions } from '../../models/EdgeData';
 
 const calculateEdgeEndPoint = (startPoint: Coord2D, direction: EdgeDirection, length: number): Coord2D => {
   switch (direction) {
